legacy/requirements.txt
transformers
torch
google-generativeai
python-dotenv
requests
streamlit
gtts